+++ platforms/core-runtime/daemon-services/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.PluginServiceRegistry
@@ -1,1 +1,1 @@
-org.gradle.launcher.daemon.DaemonServices
+org.gradle.internal.daemon.services.DaemonServices
