+++ platforms/core-runtime/daemon-services/src/test/groovy/org/gradle/internal/daemon/clientinput/StdInStreamTest.groovy
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.launcher.daemon.server.clientinput
+package org.gradle.internal.daemon.clientinput
+
 
 import org.gradle.internal.logging.events.OutputEventListener
 import org.gradle.internal.logging.events.ReadStdInEvent
