+++ testing/architecture-test/src/changes/archunit_store/internal-api-nullability.txt
@@ -3058,8 +3058,8 @@ Class <org.gradle.launcher.configuration.AllProperties> is not annotated (direct
 Class <org.gradle.launcher.configuration.BuildLayoutResult> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (BuildLayoutResult.java:0)
 Class <org.gradle.launcher.configuration.InitialProperties> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (InitialProperties.java:0)
 Class <org.gradle.launcher.daemon.DaemonExecHandleBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonExecHandleBuilder.java:0)
-Class <org.gradle.launcher.daemon.DaemonServices$DaemonGradleUserHomeServices> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonServices.java:0)
-Class <org.gradle.launcher.daemon.DaemonServices> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonServices.java:0)
+Class <org.gradle.internal.daemon.services.DaemonServices$DaemonGradleUserHomeServices> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonServices.java:0)
+Class <org.gradle.internal.daemon.services.DaemonServices> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonServices.java:0)
 Class <org.gradle.launcher.daemon.bootstrap.DaemonMain$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonMain.java:0)
 Class <org.gradle.launcher.daemon.bootstrap.DaemonMain> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonMain.java:0)
 Class <org.gradle.launcher.daemon.bootstrap.DaemonOutputConsumer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonOutputConsumer.java:0)
@@ -3177,7 +3177,7 @@ Class <org.gradle.launcher.daemon.server.api.DaemonUnavailableException> is not 
 Class <org.gradle.launcher.daemon.server.api.HandleInvalidateVirtualFileSystem> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (HandleInvalidateVirtualFileSystem.java:0)
 Class <org.gradle.launcher.daemon.server.api.HandleReportStatus> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (HandleReportStatus.java:0)
 Class <org.gradle.launcher.daemon.server.api.HandleStop> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (HandleStop.java:0)
-Class <org.gradle.launcher.daemon.server.api.StdinHandler> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (StdinHandler.java:0)
+Class <org.gradle.internal.daemon.clientinput.StdinHandler> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (StdinHandler.java:0)
 Class <org.gradle.launcher.daemon.server.exec.BuildCommandOnly> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (BuildCommandOnly.java:0)
 Class <org.gradle.launcher.daemon.server.exec.DaemonCommandExecuter> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonCommandExecuter.java:0)
 Class <org.gradle.launcher.daemon.server.exec.DaemonConnectionBackedEventConsumer$ForwardEvents> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonConnectionBackedEventConsumer.java:0)
@@ -4936,7 +4936,7 @@ Class <org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClass
 Class <org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClassLoaderRegistry$2> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ClientSidePayloadClassLoaderRegistry.java:0)
 Class <org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClassLoaderRegistry$LocalClassLoaderMapping> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ClientSidePayloadClassLoaderRegistry.java:0)
 Class <org.gradle.tooling.internal.provider.serialization.ClientSidePayloadClassLoaderRegistry> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ClientSidePayloadClassLoaderRegistry.java:0)
-Class <org.gradle.tooling.internal.provider.serialization.DaemonSidePayloadClassLoaderFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonSidePayloadClassLoaderFactory.java:0)
+Class <org.gradle.internal.daemon.serialization.DaemonSidePayloadClassLoaderFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DaemonSidePayloadClassLoaderFactory.java:0)
 Class <org.gradle.tooling.internal.provider.serialization.DefaultPayloadClassLoaderRegistry$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultPayloadClassLoaderRegistry.java:0)
 Class <org.gradle.tooling.internal.provider.serialization.DefaultPayloadClassLoaderRegistry$2> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultPayloadClassLoaderRegistry.java:0)
 Class <org.gradle.tooling.internal.provider.serialization.DefaultPayloadClassLoaderRegistry$ClassLoaderSpecVisitor> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultPayloadClassLoaderRegistry.java:0)
